Consolidate unit test setup and drop py2-compat from mocks
