+++ disk/tests/conftest.py
@@ -2,34 +2,14 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 
-import mock
 import pytest
 
-from datadog_checks.dev.utils import ON_WINDOWS, mock_context_manager
+from datadog_checks.dev.docker import using_windows_containers
+from datadog_checks.dev.utils import ON_WINDOWS
 
-from .common import E2E_METADATA
 from .metrics import CORE_COUNTS, CORE_GAUGES, CORE_RATES, UNIX_GAUGES
-from .mocks import (
-    MockDiskIOMetrics,
-    MockDiskMetrics,
-    MockInodesMetrics,
-    MockPart,
-    mock_blkid_cache_file,
-    mock_blkid_cache_file_no_label,
-)
 
-
-@pytest.fixture
-def psutil_mocks():
-    if ON_WINDOWS:
-        mock_statvfs = mock_context_manager()
-    else:
-        mock_statvfs = mock.patch('os.statvfs', return_value=MockInodesMetrics(), __name__='statvfs')
-
-    with mock.patch('psutil.disk_partitions', return_value=[MockPart()], __name__='disk_partitions'), mock.patch(
-        'psutil.disk_usage', return_value=MockDiskMetrics(), __name__='disk_usage'
-    ), mock.patch('psutil.disk_io_counters', return_value=MockDiskIOMetrics()), mock_statvfs:
-        yield
+E2E_METADATA = {'docker_platform': 'windows' if using_windows_containers() else 'linux'}
 
 
 @pytest.fixture(scope='session')
@@ -45,16 +25,6 @@ def instance_basic_volume():
 @pytest.fixture(scope='session')
 def instance_basic_mount():
     return {'use_mount': 'true', 'tag_by_label': False}
-
-
-@pytest.fixture(scope='session')
-def instance_blkid_cache_file():
-    return {'blkid_cache_file': mock_blkid_cache_file()}
-
-
-@pytest.fixture(scope='session')
-def instance_blkid_cache_file_no_label():
-    return {'blkid_cache_file': mock_blkid_cache_file_no_label()}
 
 
 @pytest.fixture(scope='session')
