+++ disk/tests/test_unit.py
@@ -1,6 +1,7 @@
 # (C) Datadog, Inc. 2018-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import os
 import re
 from itertools import chain
 
@@ -9,12 +10,107 @@ import pytest
 
 from datadog_checks.base.utils.platform import Platform
 from datadog_checks.base.utils.timeout import TimeoutException
-from datadog_checks.dev.utils import get_metadata_metrics
+from datadog_checks.dev.testing import requires_windows
+from datadog_checks.dev.utils import ON_WINDOWS, get_metadata_metrics, mock_context_manager
 from datadog_checks.disk import Disk
 from datadog_checks.disk.disk import IGNORE_CASE
 
-from .common import DEFAULT_DEVICE_BASE_NAME, DEFAULT_DEVICE_NAME, DEFAULT_FILE_SYSTEM, DEFAULT_MOUNT_POINT
-from .mocks import MockDiskMetrics, MockPart, mock_blkid_output, mock_lsblk_output
+if ON_WINDOWS:
+    DEFAULT_DEVICE_NAME = 'c:'
+    DEFAULT_DEVICE_BASE_NAME = 'c:'
+    DEFAULT_FILE_SYSTEM = 'ntfs'
+    DEFAULT_MOUNT_POINT = 'c:'
+else:
+    DEFAULT_DEVICE_NAME = '/dev/sda1'
+    DEFAULT_DEVICE_BASE_NAME = 'sda1'
+    DEFAULT_FILE_SYSTEM = 'ext4'
+    DEFAULT_MOUNT_POINT = '/'
+
+HERE = os.path.dirname(os.path.abspath(__file__))
+
+
+def mock_blkid_output():
+    """
+    Load fixtures from tests/fixtures/ folder and return a tuple matching the
+    return value of `get_subprocess_output`
+    """
+    with open(os.path.join(HERE, 'fixtures', 'blkid')) as f:
+        return f.read(), '', ''
+
+
+def mock_lsblk_output():
+    """
+    Load fixtures from tests/fixtures/ folder and return a tuple matching the
+    return value of `get_subprocess_output`
+    """
+    with open(os.path.join(HERE, 'fixtures', 'lsblk'), 'r') as f:
+        return f.read(), '', ''
+
+
+def mock_blkid_cache_file():
+    return os.path.join(HERE, 'fixtures', 'blkid_cache_file')
+
+
+def mock_blkid_cache_file_no_label():
+    return os.path.join(HERE, 'fixtures', 'blkid_cache_file_no_label')
+
+
+class MockPart:
+    def __init__(
+        self, device=DEFAULT_DEVICE_NAME, fstype=DEFAULT_FILE_SYSTEM, mountpoint=DEFAULT_MOUNT_POINT, opts='ro'
+    ):
+        self.device = device
+        self.fstype = fstype
+        self.mountpoint = mountpoint
+        self.opts = opts
+
+
+class MockDiskMetrics:
+    total = 5 * 1024
+    used = 4 * 1024
+    free = 1 * 1024
+    percent = 80
+    read_time = 50
+    write_time = 90
+
+
+class MockDiskIOMetrics(dict):
+    def __init__(self, device=DEFAULT_DEVICE_NAME):
+        super(MockDiskIOMetrics, self).__init__()
+        self[device] = MockDiskMetrics()
+
+
+class MockInodesMetrics:
+    f_files = 10
+    f_ffree = 9
+
+
+class MockIoCountersMetrics:
+    read_time = 15
+    write_time = 25
+
+
+@pytest.fixture(scope='session')
+def instance_blkid_cache_file():
+    return {'blkid_cache_file': mock_blkid_cache_file()}
+
+
+@pytest.fixture(scope='session')
+def instance_blkid_cache_file_no_label():
+    return {'blkid_cache_file': mock_blkid_cache_file_no_label()}
+
+
+@pytest.fixture
+def psutil_mocks():
+    if ON_WINDOWS:
+        mock_statvfs = mock_context_manager()
+    else:
+        mock_statvfs = mock.patch('os.statvfs', return_value=MockInodesMetrics(), __name__='statvfs')
+
+    with mock.patch('psutil.disk_partitions', return_value=[MockPart()], __name__='disk_partitions'), mock.patch(
+        'psutil.disk_usage', return_value=MockDiskMetrics(), __name__='disk_usage'
+    ), mock.patch('psutil.disk_io_counters', return_value=MockDiskIOMetrics()), mock_statvfs:
+        yield
 
 
 def test_default_options():
@@ -365,3 +461,385 @@ def test_include_all_devices(aggregator, gauge_metrics, rate_metrics, dd_run_che
     with mock.patch('psutil.disk_partitions', return_value=[]) as m:
         dd_run_check(c)
         m.assert_called_with(all=False)
+
+
+def test_default_casing():
+    if ON_WINDOWS:
+        assert IGNORE_CASE == re.I
+    else:
+        assert IGNORE_CASE == 0
+
+
+def test_default_mock():
+    c = Disk('disk', {}, [{}])
+
+    assert c.exclude_disk(MockPart()) is False
+
+
+def assert_regex_equal(a, b):
+    assert a.pattern == b.pattern
+    assert a.flags == b.flags
+
+
+def test_bad_config_string_regex():
+    instance = {
+        'file_system_include': 'test',
+        'file_system_exclude': 'test',
+        'device_include': 'test',
+        'device_exclude': 'test',
+        'mount_point_include': 'test',
+        'mount_point_exclude': 'test',
+    }
+    c = Disk('disk', {}, [instance])
+
+    assert_regex_equal(c._file_system_include, re.compile('test', re.I))
+    assert_regex_equal(c._file_system_exclude, re.compile('test|iso9660$|tracefs$', re.I))
+    assert_regex_equal(c._device_include, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._device_exclude, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._mount_point_include, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._mount_point_exclude, re.compile('test|(/host)?/proc/sys/fs/binfmt_misc$', IGNORE_CASE))
+
+
+def test_ignore_empty_regex():
+    instance = {
+        'file_system_include': ['test', ''],
+        'file_system_exclude': ['test', ''],
+        'device_include': ['test', ''],
+        'device_exclude': ['test', ''],
+        'mount_point_include': ['test', ''],
+        'mount_point_exclude': ['test', ''],
+    }
+    c = Disk('disk', {}, [instance])
+
+    assert_regex_equal(c._file_system_include, re.compile('test', re.I))
+    assert_regex_equal(c._file_system_exclude, re.compile('test|iso9660$|tracefs$', re.I))
+    assert_regex_equal(c._device_include, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._device_exclude, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._mount_point_include, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._mount_point_exclude, re.compile('test|(/host)?/proc/sys/fs/binfmt_misc$', IGNORE_CASE))
+
+
+def test_exclude_bad_devices():
+    c = Disk('disk', {}, [{}])
+
+    assert c.exclude_disk(MockPart(device='')) is True
+    assert c.exclude_disk(MockPart(device='none')) is True
+
+
+@requires_windows
+def test_exclude_cdrom():
+    c = Disk('disk', {}, [{}])
+
+    assert c.exclude_disk(MockPart(fstype='ISO9660')) is True
+    assert c.exclude_disk(MockPart(opts='rw,cdrom')) is True
+
+
+def test_file_system_include():
+    instance = {'file_system_include': ['ext[34]', 'ntfs']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(fstype='ext3')) is False
+    assert c.exclude_disk(MockPart(fstype='ext4')) is False
+    assert c.exclude_disk(MockPart(fstype='NTFS')) is False
+    assert c.exclude_disk(MockPart(fstype='apfs')) is True
+
+
+def test_file_system_exclude():
+    instance = {'file_system_exclude': ['fat']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(fstype='FAT32')) is True
+    assert c.exclude_disk(MockPart(fstype='zfs')) is False
+
+
+def test_file_system_include_exclude():
+    instance = {'file_system_include': ['ext[2-4]'], 'file_system_exclude': ['ext2']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(fstype='ext2')) is True
+    assert c.exclude_disk(MockPart(fstype='ext3')) is False
+    assert c.exclude_disk(MockPart(fstype='ext4')) is False
+    assert c.exclude_disk(MockPart(fstype='NTFS')) is True
+
+
+def test_device_include():
+    instance = {'device_include': ['/dev/sda[1-3]', 'c:']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(device='/dev/sda1')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda2')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda3')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda4')) is True
+    assert c.exclude_disk(MockPart(device='path/dev/sda1')) is True
+
+    assert c.exclude_disk(MockPart(device='c:\\')) is False
+    assert c.exclude_disk(MockPart(device='path\\c:\\')) is True
+
+
+def test_device_exclude():
+    instance = {'device_exclude': ['/dev/sda[1-3]']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(device='/dev/sda1')) is True
+    assert c.exclude_disk(MockPart(device='/dev/sda2')) is True
+    assert c.exclude_disk(MockPart(device='/dev/sda3')) is True
+    assert c.exclude_disk(MockPart(device='/dev/sda4')) is False
+
+
+def test_device_include_exclude():
+    instance = {'device_include': ['/dev/sda[1-3]'], 'device_exclude': ['/dev/sda3']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(device='/dev/sda1')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda2')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda3')) is True
+    assert c.exclude_disk(MockPart(device='/dev/sda4')) is True
+
+
+def test_mount_point_include():
+    instance = {'mount_point_include': ['/dev/sda[1-3]', 'c:']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda1')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda2')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda3')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda4')) is True
+    assert c.exclude_disk(MockPart(mountpoint='path/dev/sda1')) is True
+
+    assert c.exclude_disk(MockPart(mountpoint='c:\\')) is False
+    assert c.exclude_disk(MockPart(mountpoint='path\\c:\\')) is True
+
+
+def test_mount_point_exclude():
+    instance = {'mount_point_exclude': ['/dev/sda[1-3]']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda1')) is True
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda2')) is True
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda3')) is True
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda4')) is False
+
+
+def test_mount_point_include_exclude():
+    instance = {'mount_point_include': ['/dev/sda[1-3]'], 'mount_point_exclude': ['/dev/sda3']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda1')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda2')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda3')) is True
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda4')) is True
+
+
+def test_all_partitions_allow_no_device():
+    instance = {'all_partitions': 'true', 'mount_point_exclude': ['/run$']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(device='', mountpoint='/run')) is True
+    assert c.exclude_disk(MockPart(device='', mountpoint='/run/shm')) is False
+
+
+def test_bad_config_string_regex_deprecated():
+    instance = {
+        'file_system_whitelist': 'test',
+        'file_system_blacklist': 'test',
+        'device_whitelist': 'test',
+        'device_blacklist': 'test',
+        'mount_point_whitelist': 'test',
+        'mount_point_blacklist': 'test',
+    }
+    c = Disk('disk', {}, [instance])
+
+    assert_regex_equal(c._file_system_include, re.compile('test', re.I))
+    assert_regex_equal(c._file_system_exclude, re.compile('test|iso9660$|tracefs$', re.I))
+    assert_regex_equal(c._device_include, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._device_exclude, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._mount_point_include, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._mount_point_exclude, re.compile('test|(/host)?/proc/sys/fs/binfmt_misc$', IGNORE_CASE))
+
+
+def test_ignore_empty_regex_deprecated():
+    instance = {
+        'file_system_whitelist': ['test', ''],
+        'file_system_blacklist': ['test', ''],
+        'device_whitelist': ['test', ''],
+        'device_blacklist': ['test', ''],
+        'mount_point_whitelist': ['test', ''],
+        'mount_point_blacklist': ['test', ''],
+    }
+    c = Disk('disk', {}, [instance])
+
+    assert_regex_equal(c._file_system_include, re.compile('test', re.I))
+    assert_regex_equal(c._file_system_exclude, re.compile('test|iso9660$|tracefs$', re.I))
+    assert_regex_equal(c._device_include, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._device_exclude, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._mount_point_include, re.compile('test', IGNORE_CASE))
+    assert_regex_equal(c._mount_point_exclude, re.compile('test|(/host)?/proc/sys/fs/binfmt_misc$', IGNORE_CASE))
+
+
+def test_file_system_whitelist_deprecated():
+    instance = {'file_system_whitelist': ['ext[34]', 'ntfs']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(fstype='ext3')) is False
+    assert c.exclude_disk(MockPart(fstype='ext4')) is False
+    assert c.exclude_disk(MockPart(fstype='NTFS')) is False
+    assert c.exclude_disk(MockPart(fstype='apfs')) is True
+
+
+def test_file_system_blacklist_deprecated():
+    instance = {'file_system_blacklist': ['fat']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(fstype='FAT32')) is True
+    assert c.exclude_disk(MockPart(fstype='zfs')) is False
+
+
+def test_file_system_whitelist_blacklist_deprecated():
+    instance = {'file_system_whitelist': ['ext[2-4]'], 'file_system_blacklist': ['ext2']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(fstype='ext2')) is True
+    assert c.exclude_disk(MockPart(fstype='ext3')) is False
+    assert c.exclude_disk(MockPart(fstype='ext4')) is False
+    assert c.exclude_disk(MockPart(fstype='NTFS')) is True
+
+
+def test_device_whitelist_deprecated():
+    instance = {'device_whitelist': ['/dev/sda[1-3]', 'c:']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(device='/dev/sda1')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda2')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda3')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda4')) is True
+    assert c.exclude_disk(MockPart(device='path/dev/sda1')) is True
+
+    assert c.exclude_disk(MockPart(device='c:\\')) is False
+    assert c.exclude_disk(MockPart(device='path\\c:\\')) is True
+
+
+def test_device_blacklist_deprecated():
+    instance = {'device_blacklist': ['/dev/sda[1-3]']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(device='/dev/sda1')) is True
+    assert c.exclude_disk(MockPart(device='/dev/sda2')) is True
+    assert c.exclude_disk(MockPart(device='/dev/sda3')) is True
+    assert c.exclude_disk(MockPart(device='/dev/sda4')) is False
+
+
+def test_device_whitelist_blacklist_deprecated():
+    instance = {'device_whitelist': ['/dev/sda[1-3]'], 'device_blacklist': ['/dev/sda3']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(device='/dev/sda1')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda2')) is False
+    assert c.exclude_disk(MockPart(device='/dev/sda3')) is True
+    assert c.exclude_disk(MockPart(device='/dev/sda4')) is True
+
+
+def test_mount_point_whitelist_deprecated():
+    instance = {'mount_point_whitelist': ['/dev/sda[1-3]', 'c:']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda1')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda2')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda3')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda4')) is True
+    assert c.exclude_disk(MockPart(mountpoint='path/dev/sda1')) is True
+
+    assert c.exclude_disk(MockPart(mountpoint='c:\\')) is False
+    assert c.exclude_disk(MockPart(mountpoint='path\\c:\\')) is True
+
+
+def test_mount_point_blacklist_deprecated():
+    instance = {'mount_point_blacklist': ['/dev/sda[1-3]']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda1')) is True
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda2')) is True
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda3')) is True
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda4')) is False
+
+
+def test_mount_point_whitelist_blacklist_deprecated():
+    instance = {'mount_point_whitelist': ['/dev/sda[1-3]'], 'mount_point_blacklist': ['/dev/sda3']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda1')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda2')) is False
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda3')) is True
+    assert c.exclude_disk(MockPart(mountpoint='/dev/sda4')) is True
+
+
+def test_all_partitions_allow_no_device_deprecated():
+    instance = {'all_partitions': 'true', 'mount_point_blacklist': ['/run$']}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(device='', mountpoint='/run')) is True
+    assert c.exclude_disk(MockPart(device='', mountpoint='/run/shm')) is False
+
+
+def test_legacy_config():
+    instance = {
+        'excluded_filesystems': ['test', ''],
+        'excluded_disks': ['test1', ''],
+        'excluded_disk_re': 'test2',
+        'excluded_mountpoint_re': 'test',
+    }
+    c = Disk('disk', {}, [instance])
+
+    assert_regex_equal(c._file_system_exclude, re.compile('iso9660$|tracefs$|test$', re.I))
+    assert_regex_equal(c._device_exclude, re.compile('test1$|test2', IGNORE_CASE))
+    assert_regex_equal(c._mount_point_exclude, re.compile('(/host)?/proc/sys/fs/binfmt_misc$|test', IGNORE_CASE))
+
+
+def test_legacy_exclude_disk():
+    """
+    Test legacy exclusion logic config
+    """
+    instance = {
+        'use_mount': 'no',
+        'excluded_filesystems': ['aaaaaa'],
+        'excluded_mountpoint_re': '^/run$',
+        'excluded_disks': ['bbbbbb'],
+        'excluded_disk_re': '^tev+$',
+    }
+    c = Disk('disk', {}, [instance])
+
+    # should pass, default mock is a normal disk
+    assert c.exclude_disk(MockPart()) is False
+
+    # standard fake devices
+    assert c.exclude_disk(MockPart(device='')) is True
+    assert c.exclude_disk(MockPart(device='none')) is True
+    assert c.exclude_disk(MockPart(device='udev')) is False
+
+    # excluded filesystems list
+    assert c.exclude_disk(MockPart(fstype='aaaaaa')) is True
+    assert c.exclude_disk(MockPart(fstype='a')) is False
+
+    # excluded devices list
+    assert c.exclude_disk(MockPart(device='bbbbbb')) is True
+    assert c.exclude_disk(MockPart(device='b')) is False
+
+    # excluded devices regex
+    assert c.exclude_disk(MockPart(device='tevvv')) is True
+    assert c.exclude_disk(MockPart(device='tevvs')) is False
+
+    # and now with all_partitions
+    c._all_partitions = True
+    assert c.exclude_disk(MockPart(device='')) is False
+    assert c.exclude_disk(MockPart(device='none')) is False
+    assert c.exclude_disk(MockPart(device='udev')) is False
+    # excluded mountpoint regex
+    assert c.exclude_disk(MockPart(device='sdz', mountpoint='/run')) is True
+    assert c.exclude_disk(MockPart(device='sdz', mountpoint='/run/shm')) is False
+
+
+def test_legacy_device_exclusion_logic_no_name():
+    instance = {'use_mount': 'yes', 'excluded_mountpoint_re': '^/run$', 'all_partitions': 'yes'}
+    c = Disk('disk', {}, [instance])
+
+    assert c.exclude_disk(MockPart(device='', mountpoint='/run')) is True
+    assert c.exclude_disk(MockPart(device='', mountpoint='/run/shm')) is False
